Rename stores directory to store.
