+++ lib/index.js
@@ -1,6 +1,6 @@
 exports.state = require('./middleware/state');
 
-exports.SessionStore = require('./stores/session');
+exports.SessionStore = require('./store/session');
 
 exports.middleware = {};
 exports.middleware.clean = require('./middleware/clean');
