+++ lib/middleware/clean.js
@@ -1,4 +1,4 @@
-var SessionStore = require('../stores/session');
+var SessionStore = require('../store/session');
 
 
 module.exports = function(store, options) {
