+++ lib/middleware/state.js
@@ -8,7 +8,7 @@ var uri = require('url')
   , utils = require('../utils')
   , validate = require('../validate')
   , debug = require('debug')('flowstate')
-  , SessionStore = require('../stores/session');
+  , SessionStore = require('../store/session');
 
   // TIPS: Generally, push a state before redirecting (unless wizard style
   //  alternatively, handler can lazily push, for its own state, to preserve storage
