+++ test/integration/federated/oauth2.test.js
@@ -2,7 +2,7 @@ var chai = require('chai')
   , expect = require('chai').expect
   , sinon = require('sinon')
   , state = require('../../../lib/middleware/state')
-  , SessionStore = require('../../../lib/stores/session')
+  , SessionStore = require('../../../lib/store/session')
 
   // TODO: Move this to login/federated
 describe('integration: sso/oauth2', function() {
