+++ test/integration/oauth2.test.js
@@ -2,7 +2,7 @@ var chai = require('chai')
   , expect = require('chai').expect
   , sinon = require('sinon')
   , state = require('../../lib/middleware/state')
-  , SessionStore = require('../../lib/stores/session');
+  , SessionStore = require('../../lib/store/session');
 
 
 describe('GET /oauth2/authorize', function() {
